resolve PR comments
