+++ src/mito2/src/compaction.rs
@@ -14,6 +14,7 @@
 
 mod buckets;
 pub mod compactor;
+mod json2;
 pub mod memory_manager;
 pub mod picker;
 mod reader;
@@ -31,7 +32,7 @@ use common_meta::key::SchemaMetadataManagerRef;
 use common_telemetry::{debug, error};
 use common_time::TimeToLive;
 use common_time::range::TimestampRange;
-pub(crate) use reader::{
+pub(crate) use json2::{
     Json2RewritePlans, collect_json2_rewrite_plans, rewrite_json2_batch, rewrite_json2_schema,
 };
 pub use scheduler::CompactionRequest;
