+++ src/mito2/src/read/flat_projection.rs
@@ -38,7 +38,7 @@ use store_api::storage::ColumnId;
 use crate::cache::CacheStrategy;
 use crate::error::{InvalidRequestSnafu, RecordBatchSnafu, Result};
 use crate::read::projection::{read_column_ids_from_projection, repeated_vector_with_cache};
-use crate::read::read_columns::ReadColumns;
+use crate::read::read_columns::{JsonTargetTypes, ReadColumns};
 use crate::sst::parquet::Json2RewriteTargets;
 use crate::sst::parquet::flat_format::sst_column_id_indices;
 use crate::sst::parquet::format::FormatProjection;
@@ -128,10 +128,8 @@ impl FlatProjectionMapper {
             output_col_ids.push(col.column_id);
 
             let mut schema = col.column_schema.clone();
-            if let Some(data_type) =
-                json2_read_datatype(col.column_id, &schema.data_type, &read_cols)
-            {
-                schema.data_type = data_type;
+            if let Some(data_type) = read_cols.json_target_type(col.column_id) {
+                schema.data_type = ConcreteDataType::json2(data_type.clone());
             }
             col_schemas.push(schema);
         }
@@ -148,13 +146,8 @@ impl FlatProjectionMapper {
             read_cols.clone(),
         );
 
-        let mut batch_schema = flat_projected_columns(metadata, &format_projection);
-
-        for (column_id, data_type) in batch_schema.iter_mut() {
-            if let Some(updated) = json2_read_datatype(*column_id, data_type, &read_cols) {
-                *data_type = updated;
-            }
-        }
+        let batch_schema =
+            flat_projected_columns(metadata, &format_projection, read_cols.json_target_types());
 
         let input_arrow_schema =
             compute_input_arrow_schema(metadata, &batch_schema, &read_cols, json2_rewrite_targets);
@@ -406,20 +399,6 @@ impl FlatProjectionMapper {
     }
 }
 
-fn json2_read_datatype(
-    column_id: ColumnId,
-    data_type: &ConcreteDataType,
-    read_cols: &ReadColumns,
-) -> Option<ConcreteDataType> {
-    let json_type = data_type.as_json()?;
-    if !json_type.is_json2() {
-        return None;
-    }
-    read_cols
-        .json_target_type(column_id)
-        .map(|x| ConcreteDataType::json2(x.clone()))
-}
-
 fn single_value_string_dictionary<'a>(
     array: &'a Arc<dyn Array>,
     output_type: &ConcreteDataType,
@@ -442,12 +421,13 @@ fn single_value_string_dictionary<'a>(
     (dict_array.values().len() == 1 && dict_array.null_count() == 0).then_some(dict_array)
 }
 
-/// Returns ids and datatypes of columns of the output batch after applying the `projection`.
+/// Returns ids and datatypes of columns after applying the projection and JSON2 target types.
 ///
 /// It adds the time index column if it doesn't present in the projection.
 pub(crate) fn flat_projected_columns(
     metadata: &RegionMetadata,
     format_projection: &FormatProjection,
+    json_target_types: &JsonTargetTypes,
 ) -> Vec<(ColumnId, ConcreteDataType)> {
     let time_index = metadata.time_index_column();
     let num_columns = if format_projection
@@ -460,16 +440,18 @@ pub(crate) fn flat_projected_columns(
     };
     let mut schema = vec![None; num_columns];
     for (column_id, index) in &format_projection.column_id_to_projected_index {
-        // Safety: FormatProjection ensures the id is valid.
-        schema[*index] = Some((
-            *column_id,
+        let data_type = if let Some(json_type) = json_target_types.get(column_id) {
+            ConcreteDataType::json2(json_type.clone())
+        } else {
+            // Safety: FormatProjection ensures the id is valid.
             metadata
                 .column_by_id(*column_id)
                 .unwrap()
                 .column_schema
                 .data_type
-                .clone(),
-        ));
+                .clone()
+        };
+        schema[*index] = Some((*column_id, data_type));
     }
     if num_columns != format_projection.column_id_to_projected_index.len() {
         schema[num_columns - 1] = Some((
